float

+++ eval/evalAnomaly.py
@@ -236,7 +236,7 @@ def main():
                     # Usiamo direttamente le probabilità
                     msp_score = (1.0 - torch.max(sem_seg_probs[0], dim=0)[0]).cpu().numpy()
 
-                    p = sem_seg_probs[0]
+                    p = sem_seg_probs[0].float()
                     entropy_score = (p * torch.log(p + 1e-7)).sum(dim=0).cpu().numpy()
 
                     rba_score = calculate_rba(pixel_logits)
